Added update method.

+++ LiveDrawing/app/src/main/java/com/example/livedrawing/Particle.java
@@ -14,7 +14,7 @@ public class Particle {
         velocity.y = direction.y;
     }
 
-    public void update(Float fps) {
+    public void update(long fps) {
         position.x += velocity.x;
         position.y += velocity.y;
     }
